Start configure tool UI, add serial features, begin dynamic descriptor support

+++ firmware/EEPROM/EEPROM_IO.cpp
@@ -27,7 +27,7 @@ uint8_t EEPROM_IO::initialize() {
     #endif
 }
 
-uint8_t EEPROM_IO::read(int loc, uint8_t* buf[], size_t n=1) {
+int EEPROM_IO::read(int loc, uint8_t* buf[], size_t n=1) {
     #ifdef EEPROM_EXTERNAL
         return ee.read(loc, buf, n);
     #else
@@ -45,7 +45,7 @@ uint8_t EEPROM_IO::readByte(int loc) {
     #endif
 }
 
-uint8_t EEPROM_IO::write(int loc, uint8_t* buf[], size_t n) {
+int EEPROM_IO::write(int loc, uint8_t* buf[], size_t n) {
     #ifdef EEPROM_EXTERNAL
         return ee.write(loc, buf, n);
     #else
@@ -55,7 +55,7 @@ uint8_t EEPROM_IO::write(int loc, uint8_t* buf[], size_t n) {
     #endif
 }
 
-uint8_t EEPROM_IO::update(int loc, uint8_t* buf[], size_t n) {
+int EEPROM_IO::update(int loc, uint8_t* buf[], size_t n) {
     #ifdef EEPROM_EXTERNAL
         // probably have some loop here to check first
         return ee.write(loc, buf, n);
@@ -66,11 +66,11 @@ uint8_t EEPROM_IO::update(int loc, uint8_t* buf[], size_t n) {
     #endif
 }
 
-uint8_t EEPROM_IO::readConfig(uint8_t* config) {
+int EEPROM_IO::readConfig(uint8_t* config) {
     return this->read(0, config, sizeof(config));
 }
 
-uint8_t EEPROM_IO::updateConfig(uint8_t* config) {
+int EEPROM_IO::updateConfig(uint8_t* config) {
     return this->update(0, config, sizeof(config));
 }
